chore: regen schema move api files into their own directory
