+++ src/api/gql/index.ts
@@ -1,0 +1,3 @@
+export { LoginGQL } from "./loginMutation";
+export { MeGQL } from "./meQuery";
+export * from "./__generated__";
