+++ src/api/gql/meQuery.ts
@@ -1,0 +1,18 @@
+import gql from "graphql-tag";
+
+export const MeGQL = gql`
+  query Me($userId: ID!) {
+    me(userId: $userId) {
+      ... on User {
+        email
+        campaigns {
+          id
+          name
+        }
+      }
+      ... on UserNotFound {
+        message
+      }
+    }
+  }
+`;
