+++ src/api/index.ts
@@ -1,0 +1,2 @@
+export { auth } from "./auth";
+export { me } from "./me";
